Pull tag comparator into utils
